+++ src/lib.rs
@@ -1,5 +1,4 @@
 mod num_traits;
-mod range;
 mod traits;
 mod vec2;
 mod vec2_range_iter;
@@ -8,7 +7,7 @@ mod vec3_range_iter;
 mod vec4;
 
 pub use self::{
-    num_traits::*, range::*, traits::*, vec2::*, vec2_range_iter::*, vec3::*,
+    num_traits::*, traits::*, vec2::*, vec2_range_iter::*, vec3::*,
     vec3_range_iter::*, vec4::*,
 };