+++ src/traits/mod.rs
@@ -1,4 +1,5 @@
+mod cast_ext;
 mod range_ext;
 mod two_component;
 
-pub use self::{range_ext::*, two_component::*};
+pub use self::{cast_ext::*, range_ext::*, two_component::*};